refactor: split tests into dedicated parts, optimize test exec time
