+++ .github/workflows/ci.yml
@@ -20,8 +20,8 @@ jobs:
       - name: Lint
         run: npm run lint
 
-  Unit-Test:
-    name: ${{ matrix.os }} - Unit Tests
+  VSCode-Test:
+    name: ${{ matrix.os }} - VS Code Tests
     needs: Lint
     runs-on: ${{ matrix.os }}
     strategy:
@@ -29,20 +29,36 @@ jobs:
       matrix:
         os: [ ubuntu-latest, macos-latest, windows-latest ]
     steps:
-      - uses: actions/checkout@v3
-      - uses: actions/setup-node@v3
+      - uses: actions/checkout@v4
+      - uses: actions/setup-node@v4
         with:
           node-version: '24.x'
       - name: Install NPM dependencies
         run: npm ci
-      - name: Run Tests
-        uses: GabrielBB/xvfb-action@v1.0
+      - name: Run VS Code Tests on Linux
+        if: runner.os == 'Linux'
+        run: xvfb-run -a npm run test:vscode
+      - name: Run VS Code Tests
+        if: runner.os != 'Linux'
+        run: npm run test:vscode
+
+  Coverage:
+    name: Ubuntu - Coverage
+    needs: Lint
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+      - uses: actions/setup-node@v4
         with:
-          run: npm run test
+          node-version: '24.x'
+      - name: Install NPM dependencies
+        run: npm ci
+      - name: Run Coverage
+        run: xvfb-run -a npm run test:coverage
 
   Package:
     runs-on: ubuntu-latest
-    needs: [Lint, Unit-Test]
+    needs: [Lint, VSCode-Test, Coverage]
     steps:
       - uses: actions/checkout@v4
         with:
