+++ src/test/vscode/suite/providers/completion/style.test.ts
@@ -1,6 +1,5 @@
 /* eslint no-template-curly-in-string: off */
 import { expect } from 'chai';
-import { after, before, describe, it } from 'mocha';
 import * as sinon from 'sinon';
 import { Project } from '../../../../../project';
 import * as vscode from 'vscode';
